+++ packages/stentor-guards/src/index.ts
@@ -17,3 +17,4 @@ export * from "./isSchedulableResponse";
 export * from "./isScheduled";
 export * from "./isTemplatedList";
 export * from "./isTimeContextual";
+export * from "./request";
