+++ packages/stentor-handler-delegating/package.json
@@ -24,7 +24,7 @@
   },
   "dependencies": {
     "stentor-handler": "1.54.11",
-    "stentor-request": "1.54.11"
+    "stentor-utils": "1.54.11"
   },
   "peerDependencies": {
     "stentor-models": "1.x"
