+++ packages/stentor-handler-manager/package.json
@@ -26,6 +26,7 @@
     "sinon": "13.0.1",
     "sinon-chai": "3.7.0",
     "stentor-models": "1.54.11",
+    "stentor-request": "1.54.11",
     "ts-node": "10.7.0",
     "typescript": "4.6.3"
   },
@@ -34,7 +35,6 @@
     "stentor-handler": "1.54.11",
     "stentor-handler-factory": "1.54.11",
     "stentor-logger": "1.54.11",
-    "stentor-request": "1.54.11",
     "stentor-response": "1.54.11",
     "stentor-storage": "1.54.11"
   },
