+++ packages/stentor-handler-manager/src/__test__/HandlerManager.test.ts
@@ -5,10 +5,10 @@ import * as sinon from "sinon";
 import * as sinonChai from "sinon-chai";
 
 import { CONVERSATION_HANDLER_TYPE } from "stentor-constants";
-import { isHandler } from "stentor-guards";
+import { isHandler, isIntentRequest } from "stentor-guards";
 import { HandlerFactory } from "stentor-handler-factory";
 import { Context, Device, Handler, HandlerService, Request, Storage, InputUnknownRequest } from "stentor-models";
-import { IntentRequestBuilder, isIntentRequest, InputUnknownRequestBuilder } from "stentor-request";
+import { IntentRequestBuilder, InputUnknownRequestBuilder } from "stentor-request";
 import { ResponseBuilder } from "stentor-response";
 import { HandlerManager } from "../HandlerManager";
