+++ packages/stentor-handler/src/AbstractHandler/__test__/Handler.test.ts
@@ -4,7 +4,7 @@ import * as sinon from "sinon";
 import * as sinonChai from "sinon-chai";
 
 import { ContextBuilder } from "stentor-context";
-import { BASE_HANDLER_TYPE, OPTION_SELECT_ID, } from "stentor-constants";
+import { BASE_HANDLER_TYPE, OPTION_SELECT_ID, INTENT_REQUEST_TYPE } from "stentor-constants";
 import { HELP_INTENT, STOP_INTENT } from "stentor-interaction-model";
 import {
     Content,
@@ -20,7 +20,6 @@ import {
 } from "stentor-models";
 import {
     InputUnknownRequestBuilder,
-    INTENT_REQUEST_TYPE,
     IntentRequestBuilder,
     OptionSelectBuilder
 } from "stentor-request";
