+++ packages/stentor-models/src/Knowledgebase/index.ts
@@ -1,2 +1,3 @@
 /*! Copyright (c) 2022, XAPPmedia */
 export * from "./KnowledgebaseData";
+export * from "./KnowledgeBaseConfig";
