+++ packages/stentor-request/src/index.ts
@@ -1,11 +1,6 @@
 /*! Copyright (c) 2019, XAPPmedia */
 export * from "./Builders";
-export * from "./Constants";
 export * from "./findRequestDependentMatch";
 export * from "./findSystemDependentMatch";
-export * from "./getSlots";
-export * from "./Guards";
-export * from "./hasSlots";
-export * from "./keyFromRequest";
 export * from "./RequestConditionalCheck";
 export * from "./SystemConditionalCheck";
