+++ packages/stentor-utils/src/request/index.ts
@@ -1,5 +1,7 @@
 /*! Copyright (c) 2020, XAPPmedia */
 export * from "./combineRequestSlots";
+export * from "./getSlots";
 export * from "./getSlotValue";
+export * from "./hasSlots";
 export * from "./requestSlotsToString";
 export * from "./requestSlotValueToString";
